+++ service/globals.go
@@ -6,7 +6,7 @@ import (
 )
 
 // Generate Random number
-func getRand(max int64) (int, error) {
-    rand, err := rand.Int(rand.Reader, big.NewInt(max))
+func getRand(max int) (int, error) {
+    rand, err := rand.Int(rand.Reader, big.NewInt(int64(max)))
     return int(rand.Int64()), err
 }
